+++ packages/astro/src/assets/fonts/logic/normalize-remote-font-faces.ts
@@ -1,0 +1,42 @@
+import type * as unifont from 'unifont';
+import type { UrlProxy } from '../definitions.js';
+
+export function normalizeRemoteFontFaces({
+	fonts,
+	urlProxy,
+}: {
+	fonts: Array<unifont.FontFaceData>;
+	urlProxy: UrlProxy;
+}): Array<unifont.FontFaceData> {
+	return (
+		fonts
+			// Avoid getting too much font files
+			.filter((font) => (typeof font.meta?.priority === 'number' ? font.meta.priority === 0 : true))
+			// Collect URLs
+			.map((font) => {
+				// The index keeps track of encountered URLs. We can't use the index on font.src.map
+				// below because it may contain sources without urls, which would prevent preloading completely
+				let index = 0;
+				return {
+					...font,
+					src: font.src.map((source) => {
+						if ('name' in source) {
+							return source;
+						}
+						const proxied = {
+							...source,
+							originalURL: source.url,
+							url: urlProxy.proxy({
+								url: source.url,
+								// We only collect the first URL to avoid preloading fallback sources (eg. we only
+								// preload woff2 if woff is available)
+								collectPreload: index === 0,
+							}),
+						};
+						index++;
+						return proxied;
+					}),
+				};
+			})
+	);
+}
